+++ src/ContainerInstance/ContainerInstance.Autorest/custom/New-AzContainerGroupVolumeObject.ps1
@@ -21,12 +21,12 @@ Create an in-memory object for Volume.
 Create an in-memory object for Volume.
 
 .Outputs
-Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20221001Preview.Volume
+Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20240501Preview.Volume
 .Link
 https://learn.microsoft.com/powershell/module/az.ContainerInstance/new-AzContainerGroupVolumeObject
 #>
 function New-AzContainerGroupVolumeObject {
-    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20221001Preview.Volume')]
+    [OutputType('Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20240501Preview.Volume')]
     [CmdletBinding(PositionalBinding=$false)]
     Param(
 
@@ -43,7 +43,7 @@ function New-AzContainerGroupVolumeObject {
         [string]
         $AzureFileStorageAccountName,
         [Parameter(HelpMessage="The empty directory volume.")]
-        [Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20221001Preview.IVolumeEmptyDir]
+        [Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20240501Preview.IVolumeEmptyDir]
         $EmptyDir,
         [Parameter(HelpMessage="Target directory name. Must not contain or start with '..'.  If '.' is supplied, the volume directory will be the git repository.  Otherwise, if specified, the volume will contain the git repository in the subdirectory with the given name.")]
         [string]
@@ -58,12 +58,12 @@ function New-AzContainerGroupVolumeObject {
         [string]
         $Name,
         [Parameter(HelpMessage="The secret volume.")]
-        [Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20221001Preview.ISecretVolume]
+        [Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20240501Preview.ISecretVolume]
         $Secret
     )
 
     process {
-        $Object = [Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20221001Preview.Volume]::New()
+        $Object = [Microsoft.Azure.PowerShell.Cmdlets.ContainerInstance.Models.Api20240501Preview.Volume]::New()
 
         if ($PSBoundParameters.ContainsKey('AzureFileReadOnly')) {
             $Object.AzureFileReadOnly = $AzureFileReadOnly