+++ src/HDInsight/HDInsight/ChangeLog.md
@@ -20,6 +20,9 @@
 
 ## Upcoming Release
 
+## Version 6.3.0
+* Changed the type of parameter `-IdentityId` in command `Update-AzHDInsightCluster` from `string`  to `string[]`.
+
 ## Version 6.2.1
 * Fixed a bug: Error occurs when setting the same assigned identity for storage and esp configurations.